Initial draft of very simple elm chat
This version does not interact with any server yet. It just enables the
user to type their message into an input box and pseudo-send it. "Pseudo"
means the message of the user will just be appended to a list of previously
sent messages, which will be displayed above the input box. As of now,
there is no error handling for blank messages.

To model this, the 'Model' contains the current ownMessage and a
'List Message' where 'Message' is just a simple type alias. This will
come in handy later for expanding the 'Message' with additional
data (i.e. username). Now, the 'Msg' contains just two variants. One is
for updating the text in the input box and the other is just for
"broadcasting" or "sending" the actual change.

You should be able to try it out just using elm-reactor.