use App.program for side effects
With the App.beginnerProgram one is only able to communicate inside the
program itself, but not with the outside world. To mitigate that, there
is another function named 'App.program', which enables to subscribe to
changes, which happen outside of Elm and route them inside our own
update function, with a Message we know how to interpret. This happens
inside the 'subscriptions' function. Take a look at some commits further
down the line to see an example of how to do that.

Besides the new 'subscriptions' function the change from
'beginnerProgram' to 'program' entails to change 'update' and 'init' to
a function, which returns a tuple of the new Model and a 'Cmd Msg'. That
is to say a command of our own message. This means, we can send
something to the outside world and the resulting value of that call will
be another messages, which will be routed into our own app.