add simple websocket interaction with broadcast
After implementing the plumbing in previous commits this is just a
matter of installing the correct library, as can be seen in
the change for elm-package.json and wiring it up to the app.

The two main functions used to do the latter are mainly 'listen' and
'send' from the aforementioned WebSocket package. By the way, this one
is a real cracker, since it automatically handles disconnects with an
exponential backoff strategy and queues sent messages, to push them out,
when the connection is back.

Now, listening to a server will be done by adding the 'listen' function
with the serverUrl one wants to listen to and a function, which converts
the incoming string message into an actual message the app can
understand, into the 'subscriptions' function. So whenever a message
comes in from the given server, it will be converted to a message the
app can understand and routed into the update function, to get a new
state/model and rerender what has changed to the screen.

Actually sending messages is just a matter of extending our previously
defined handling of a 'Broadcast' message and as a command to the Elm
runtime, we are telling it "Hey, send this message, when our user clicks
on 'send'". The difference between this and sending it like in an
imperative language is quite a fine one. Instead of saying "do it now
and then carry on with some other stuff", we issue a command to the Elm
runtime to execute the command when it gets to it. Usually this will be
right away, but the difference is basically, instead of doing it, we
just tell, that we eventually would like a message to be sent.