use programWithFlags for environment params
Instead of using 'program', one can use 'programWithFlags' for providing
an additional configuration of the application.

In this case we are providing the url of the broadcast server to the
program. This is just a String, but a more sophisticated configuration
type would be possible too.

type alias AppConfig =
  { serverUrl : WebSocketUrl
  , name : String
  }

main : Program AppConfig Model Msg

Beaware though, that no sum types (so no 'type PersonTyp = A | B') can
be provided as a configuration. For those cases, we need to provide a
custom JSON decoder and decode the incoming String for ourselves.